+++ src/index.ts
@@ -67,8 +67,10 @@ import { extendViemChain } from './utils/extendViemChain';
 import { formatToken } from './utils/tokens';
 
 export * from './chains';
+export * from './errors';
 export * from './types/zap';
 export * from './zap/constants';
+export type { DZapClientOptions } from './dZapClient';
 export {
   AllowanceTypes,
   ApiRpcResponse,